CU-866an2bc6 | Implement "Your Albums" empty and filled state components
