+++ android/app-newm/src/main/java/io/newm/navigation/Navigation.kt
@@ -66,7 +66,8 @@ private fun NavGraphBuilder.addLibraryTree() {
         composable(Screen.LibraryLanding.route) {
             LibraryScreen(
                 onSongView = {}, //TODO: Implement on song view
-                onArtistViewDetails = {}//TODO: Implement on artist view
+                onArtistViewDetails = {},//TODO: Implement on artist view
+                onAlbumViewDetails = {}//TODO: Implement on album view
             )
         }
     }
