+++ android/core-resources/src/main/res/values/strings.xml
@@ -59,6 +59,7 @@
     <string name="title_saved_song_list">SAVED SONGS</string>
     <string name="library_empty_state_label">Nothing here yet!</string>
     <string name="library_artists">YOUR ARTISTS</string>
+    <string name="library_albums">YOUR ALBUMS</string>
 
 
 </resources>
